+++ packages/elements/emoji/src/__tests__/unit/components/picker/UploadingEmojiPickerSpec.tsx
@@ -14,7 +14,8 @@ import {
   customCategory,
   customTitle,
   userCustomTitle,
-} from '../../../../constants';
+} from '../../../../util/constants';
+import { EmojiDescription } from '../../../../types';
 import * as ImageUtil from '../../../../util/image';
 import {
   createPngFile,
@@ -39,7 +40,7 @@ import {
   uploadConfirmButton,
   uploadFailedEvent,
   uploadSucceededEvent,
-} from '../../../../analytics';
+} from '../../../../util/analytics';
 
 describe('<UploadingEmojiPicker />', () => {
   let onEvent: jest.SpyInstance;
@@ -56,7 +57,7 @@ describe('<UploadingEmojiPicker />', () => {
     const uploadPreviewEmoji = uploadPreview.find(Emoji);
     // Should show two emoji in EmojiUploadPrevew
     expect(uploadPreviewEmoji).toHaveLength(2);
-    let emoji = uploadPreviewEmoji.at(0).prop('emoji');
+    const emoji: EmojiDescription = uploadPreviewEmoji.at(0).prop('emoji');
     expect(emoji.shortName).toEqual(':cheese_burger:');
     expect((emoji.representation as any).imagePath).toEqual(pngDataURL);
   };